data/pyproject.toml
[project]
name = "data"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "holidays>=0.56",
    "lxml>=6.0.4",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "requests>=2.33.1",
    "scipy>=1.17.1",
]
